template-python-outgoing-notifications-v1/requirements.txt
azure-common==1.1.27
azure-core==1.15.0
azure-servicebus==7.3.0
certifi==2021.5.30
chardet==4.0.0
idna==2.10
isodate==0.6.0
msrest==0.6.21
oauthlib==3.1.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
uamqp==1.4.0
urllib3==1.26.5
